+++ examples/basic.rs
@@ -27,7 +27,11 @@ fn main() {
     let mut context = Context::new();
     context.add("username", &"Bob");
     context.add("numbers", &vec![1,2,3]);
+    context.add("show_all", &false);
     context.add("bio", &"<script>alert('pwnd');</script>");
+
+    // A one off template
+    Tera::one_off("hello", Context::new(), true).unwrap();
 
     match TEMPLATES.render("users/profile.html", context) {
         Ok(s) => println!("{:?}", s),